requirements.txt
asgiref==3.3.1
cffi==1.14.5
click==7.1.2
cryptography==3.4.7
Django==3.1.7
django-cors-headers==3.7.0
django-rest-knox==4.1.0
djangorestframework==3.12.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Pillow==8.1.2
pycparser==2.20
pytz==2021.1
sqlparse==0.4.1
Werkzeug==1.0.1
